+++ crates/cheatnet/src/predeployment/mod.rs
@@ -1,1 +1,3 @@
+pub mod eth;
+pub mod predeployed_contract;
 pub mod strk;